ref: brakeEfficiency float

+++ src/main/java/cam72cam/immersiverailroading/entity/Locomotive.java
@@ -533,7 +533,7 @@ public abstract class Locomotive extends FreightTank {
 	}
 
 	@Override
-	public double getBrakeSystemEfficiency() {
+	public float getBrakeSystemEfficiency() {
 		if (cogging) {
 			return 10;
 		}
